move to __init__
